Fixed return values.

+++ endianness/endianness.h
@@ -30,8 +30,8 @@ void tolitend32(uint8_t* buffer, uint32_t n);
 void tolitend16(uint8_t* buffer, uint16_t n);
 uint32_t fromlitend32(uint8_t* buffer);
 uint32_t frombigend32(uint8_t* buffer);
-uint32_t fromlitend16(uint8_t* buffer);
-uint32_t frombigend16(uint8_t* buffer);
+uint16_t fromlitend16(uint8_t* buffer);
+uint16_t frombigend16(uint8_t* buffer);
 
 #ifdef __cplusplus
 }
